+++ NEWS.md
@@ -1,12 +1,13 @@
 # rotsee UNRELEASED
 ## added
-* examples for the functions `rc_calculate_bd`, `rc_calculate_B_C_OF`, `rc_extend_crops`, `rc_extend_amendments`
+* examples for the functions `rc_calculate_bd`, `rc_calculate_bcof`, `rc_extend_crops`, `rc_extend_amendments`
 * unit tests for rc_visualize_plot
 * Vignettes to add overview of rotsee and a description of the helper functions
 
 ## changed
 * removed intermediate columns dens.sand, dens.clay, and cf as output from `rc_calculate_bd`
 * set default of initialize in `parms` to FALSE
+* updated function name `rc_calculate_B_C_OF` to `rc_calculate_bcof`.
 
 # rotsee v0.1.8 2025-11-12
 ## added